# Handle corner case of sad64x64 comparison