--- a/Source/Lib/ASM_AVX2/EbComputeSAD_Intrinsic_AVX2.c
+++ b/Source/Lib/ASM_AVX2/EbComputeSAD_Intrinsic_AVX2.c
@@ -2096,32 +2096,28 @@ void GetEightHorizontalSearchPointResults_32x32_64x64_PU_AVX2_INTRIN(
     sad_0 = _mm_slli_epi32(sad_0, 1);
     sad_1 = _mm_slli_epi32(sad_1, 1);
 
-    EB_BOOL sadChanged = EB_FALSE;
-    bestMV64x64 = 0;
+
+    bestMV64x64 = 0xff;
 
     //sad_0
     temSum = _mm_extract_epi32(sad_0, 0);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
-        sadChanged = EB_TRUE;
     }
     temSum = _mm_extract_epi32(sad_0, 1);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 1 * 4;
-        sadChanged = EB_TRUE;
     }
     temSum = _mm_extract_epi32(sad_0, 2);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 2 * 4;
-        sadChanged = EB_TRUE;
     }
     temSum = _mm_extract_epi32(sad_0, 3);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 3 * 4;
-        sadChanged = EB_TRUE;
     }
 
     //sad_1
@@ -2129,28 +2125,24 @@ void GetEightHorizontalSearchPointResults_32x32_64x64_PU_AVX2_INTRIN(
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 4 * 4;
-        sadChanged = EB_TRUE;
     }
     temSum = _mm_extract_epi32(sad_1, 1);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 5 * 4;
-        sadChanged = EB_TRUE;
     }
     temSum = _mm_extract_epi32(sad_1, 2);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 6 * 4;
-        sadChanged = EB_TRUE;
     }
     temSum = _mm_extract_epi32(sad_1, 3);
     if (temSum <= pBestSad64x64[0]) {
         pBestSad64x64[0] = temSum;
         bestMV64x64 = 7 * 4;
-        sadChanged = EB_TRUE;
     }
 
-    if (sadChanged) {
+    if (bestMV64x64 != 0xff) {
         xMv = _MVXT(mv) + (EB_S16)bestMV64x64;  yMv = _MVYT(mv);
         pBestMV64x64[0] = ((EB_U16)yMv << 16) | ((EB_U16)xMv);
     }