Changed the check for pBestSad64x64 to also
include being equal to the value, as the computed
value might equal pBestSad64x64 without a correct
corresponding motion vector ever being set.

We are thus changing to <= to include this corner
case.

Signed-off-by: Mark Feldman <mark.feldman@intel.com>

Closes #481