@tianjunwork 2020-03-19
Hi @intelmark , could you add the test result with `-asm 0`, `-asm 1` with avx2/avx512 here?
Please also check MD5. Thank you.