@intelmark 2020-03-19
Testing shows that with -asm 0 and -asm 1, both with avx2 and avx512 encodings with the latest commits in this PR are binary identical.