@intelmark 2020-03-21
Isn’t needed. See the comments in the commit


On Mar 20, 2020, at 9:37 PM, Jun <notifications@github.com> wrote:

﻿

@tianjunwork commented on this pull request.

________________________________

In Source/Lib/ASM_AVX2/EbComputeSAD_Intrinsic_AVX2.c<https://github.com/OpenVisualCloud/SVT-HEVC/pull/500#discussion_r395960848>:

> @@ -2139,11 +2139,11 @@ void GetEightHorizontalSearchPointResults_32x32_64x64_PU_AVX2_INTRIN(
         bestSad64x64 = temSum;
         bestMV64x64 = 7 * 4;
     }
-    if (bestSad64x64 <= pBestSad64x64[0]) {
-        pBestSad64x64[0] = bestSad64x64;
-        xMv = _MVXT(mv) + (EB_S16)bestMV64x64;  yMv = _MVYT(mv);
-        pBestMV64x64[0] = ((EB_U16)yMv << 16) | ((EB_U16)xMv);
-    }
+
+    pBestSad64x64[0] = bestSad64x64;


why removed the if?

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub<https://github.com/OpenVisualCloud/SVT-HEVC/pull/500#pullrequestreview-378891769>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/AMC5UQWVOGLXLJK545GZQ53RIQ76XANCNFSM4LO3M5RA>.