@tianjunwork 2020-03-25
Hi @intelmark , please rerun the CI fast type. And also check for perf. Thank you.