@tszumski 2020-03-26
Hi Mark,
Changes looks good but I have different proposition for AVX changes (less code):
https://github.com/tszumski/SVT-HEVC/commit/80362a5b909dcce6577145b98881ecbd02118177