@intelmark 2020-03-26
python SVT-HEVC_FunctionalTests.py Fast -type asm_test

Test Begin...
Total Number of Tests: 108
Total Passed: 108
Percentage Passed: 100.0%
Time Elapsed: 9 minute(s), 53 second(s) 

Speed check:
(with change)
Bin/Release/SvtHevcEncApp -i /testfiles/bbb_1920x1080_420p.yuv -w 1920 -h 1080 -asm 0 -encMode 0 -intra-period 100 -scd 0 -rc 0 -q 22 -n 100 -b Slibbb.265
Average Speed:          2.11 fps
Total Encoding Time:    47446 ms
Total Execution Time:   47952 ms
Average Latency:        33999 ms
Max Latency:            46614 ms

(master branch)
Average Speed:          2.08 fps
Total Encoding Time:    48028 ms
Total Execution Time:   48529 ms
Average Latency:        34168 ms
Max Latency:            47832 ms